+++ apps/routing_settings.go
@@ -2,21 +2,26 @@ package apps
 
 import (
 	"context"
+	"net/netip"
+	"os"
+
+	"github.com/olekukonko/tablewriter"
 
 	"github.com/Scalingo/cli/config"
 	"github.com/Scalingo/cli/io"
+	"github.com/Scalingo/go-scalingo/v11"
 	"github.com/Scalingo/go-utils/errors/v3"
 )
 
 func ForceHTTPS(ctx context.Context, appName string, enable bool) error {
 	c, err := config.ScalingoClient(ctx)
 	if err != nil {
-		return errors.Wrapf(ctx, err, "fail to get Scalingo client")
+		return errors.Wrapf(ctx, err, "get Scalingo client")
 	}
 
 	_, err = c.AppsForceHTTPS(ctx, appName, enable)
 	if err != nil {
-		return errors.Wrapf(ctx, err, "fail to configure force-https feature")
+		return errors.Wrapf(ctx, err, "configure force-https feature")
 	}
 
 	var action string
@@ -33,11 +38,11 @@ func ForceHTTPS(ctx context.Context, appName string, enable bool) error {
 func StickySession(ctx context.Context, appName string, enable bool) error {
 	c, err := config.ScalingoClient(ctx)
 	if err != nil {
-		return errors.Wrapf(ctx, err, "fail to get Scalingo client")
+		return errors.Wrapf(ctx, err, "get Scalingo client")
 	}
 	_, err = c.AppsStickySession(ctx, appName, enable)
 	if err != nil {
-		return errors.Wrapf(ctx, err, "fail to configure sticky-session feature")
+		return errors.Wrapf(ctx, err, "configure sticky-session feature")
 	}
 
 	var action string
@@ -54,12 +59,12 @@ func StickySession(ctx context.Context, appName string, enable bool) error {
 func RouterLogs(ctx context.Context, appName string, enable bool) error {
 	c, err := config.ScalingoClient(ctx)
 	if err != nil {
-		return errors.Wrapf(ctx, err, "fail to get Scalingo client")
+		return errors.Wrapf(ctx, err, "get Scalingo client")
 	}
 
 	_, err = c.AppsRouterLogs(ctx, appName, enable)
 	if err != nil {
-		return errors.Wrapf(ctx, err, "fail to configure router-logs feature")
+		return errors.Wrapf(ctx, err, "configure router-logs feature")
 	}
 
 	var action string
@@ -70,5 +75,79 @@ func RouterLogs(ctx context.Context, appName string, enable bool) error {
 	}
 
 	io.Statusf("Router logs have been %sd on %s\n", action, appName)
+	return nil
+}
+
+func AppFirewallRulesList(ctx context.Context, appName string) error {
+	c, err := config.ScalingoClient(ctx)
+	if err != nil {
+		return errors.Wrapf(ctx, err, "get Scalingo client")
+	}
+
+	rules, err := c.AppsFirewallRulesList(ctx, appName)
+	if err != nil {
+		return errors.Wrapf(ctx, err, "list app firewall rules")
+	}
+
+	if len(rules) == 0 {
+		io.Statusf("No app firewall rules configured on %s\n", appName)
+		return nil
+	}
+
+	t := tablewriter.NewWriter(os.Stdout)
+	t.Header([]string{"ID", "CIDR", "Label"})
+	for _, rule := range rules {
+		_ = t.Append([]string{rule.ID, rule.CIDR, rule.Label})
+	}
+	_ = t.Render()
+	return nil
+}
+
+func AppFirewallRuleAdd(ctx context.Context, appName string, cidr string, label string) error {
+	err := validateIPv4CIDR(ctx, cidr)
+	if err != nil {
+		return err
+	}
+
+	c, err := config.ScalingoClient(ctx)
+	if err != nil {
+		return errors.Wrapf(ctx, err, "get Scalingo client")
+	}
+
+	rule, err := c.AppsFirewallRuleCreate(ctx, appName, scalingo.AppFirewallRuleParams{
+		CIDR:  cidr,
+		Label: label,
+	})
+	if err != nil {
+		return errors.Wrapf(ctx, err, "add app firewall rule")
+	}
+
+	io.Statusf("app firewall rule %s has been added to %s\n", rule.ID, appName)
+	return nil
+}
+
+func AppFirewallRuleRemove(ctx context.Context, appName string, ruleID string) error {
+	c, err := config.ScalingoClient(ctx)
+	if err != nil {
+		return errors.Wrapf(ctx, err, "get Scalingo client")
+	}
+
+	err = c.AppsFirewallRuleDelete(ctx, appName, ruleID)
+	if err != nil {
+		return errors.Wrapf(ctx, err, "remove app firewall rule")
+	}
+
+	io.Statusf("app firewall rule %s has been removed from %s\n", ruleID, appName)
+	return nil
+}
+
+func validateIPv4CIDR(ctx context.Context, cidr string) error {
+	prefix, err := netip.ParsePrefix(cidr)
+	if err != nil {
+		return errors.Newf(ctx, "invalid IPv4 CIDR %q", cidr)
+	}
+	if !prefix.Addr().Is4() {
+		return errors.Newf(ctx, "invalid IPv4 CIDR %q: IPv6 prefixes are not supported", cidr)
+	}
 	return nil
 }