+++ cmd/commands.go
@@ -355,6 +355,11 @@ var (
 		&databaseNetPeeringsAddCommand,
 		&databaseNetPeeringsRemoveCommand,
 		&databaseNetworkConfigurationShowCommand,
+
+		// App firewall rules
+		&appFirewallRulesCommand,
+		&appFirewallRuleAddCommand,
+		&appFirewallRuleRemoveCommand,
 	}
 
 	globalCommands = []*cli.Command{